add prefix config thru .env #6
